requirements.txt
streamlit>=1.25.0
openai>=0.27.0
faiss-cpu>=1.7.4
pandas>=1.5.0
numpy>=1.23.0
sentence-transformers>=2.2.2
scikit-learn>=1.2.2
sqlite-utils>=3.13
python-dotenv>=1.0.0
Pillow>=9.5.0
requests>=2.31.0
